perf: Add optional auth decorator
