--- backend/src/core/sessions/admin_authorization/authorization_admin-core_sessions.resolver.ts
@@ -1,18 +1,0 @@
-import { Context, Query, Resolver } from '@nestjs/graphql';
-
-import { AuthorizationAdminCoreSessionsService } from './authorization_admin_authorization-core_sessions.service';
-import { AuthorizationAdminCoreSessionsObj } from './dto/authorization_admin-core_sessions.obj';
-
-import { Ctx } from '@/types/context.type';
-
-@Resolver()
-export class AuthorizationAdminCoreSessionsResolver {
-  constructor(private readonly service: AuthorizationAdminCoreSessionsService) {}
-
-  @Query(() => AuthorizationAdminCoreSessionsObj)
-  async authorization_admin_core_sessions(
-    @Context() context: Ctx
-  ): Promise<AuthorizationAdminCoreSessionsObj> {
-    return await this.service.authorization(context);
-  }
-}
