+++ backend/src/core/sessions/authorization/authorization-core_sessions.resolver.ts
@@ -1,16 +1,18 @@
 import { Context, Query, Resolver } from '@nestjs/graphql';
 
 import { AuthorizationCoreSessionsService } from './authorization-core_sessions.service';
+import { AuthorizationCoreSessionsObj } from './dto/authorization-core_sessions.obj';
 
 import { Ctx } from '@/types/context.type';
-import { User } from '@/utils/decorators/user.decorator';
 
 @Resolver()
 export class AuthorizationCoreSessionsResolver {
   constructor(private readonly service: AuthorizationCoreSessionsService) {}
 
-  @Query(() => User)
-  async authorization_core_sessions(@Context() context: Ctx): Promise<User> {
+  @Query(() => AuthorizationCoreSessionsObj)
+  async authorization_core_sessions(
+    @Context() context: Ctx
+  ): Promise<AuthorizationCoreSessionsObj> {
     return await this.service.authorization(context);
   }
 }
