+++ backend/types/context.type.ts
@@ -1,6 +1,12 @@
 import { Request, Response } from 'express';
 
+import { User } from '@/utils/decorators/user.decorator';
+
+export interface AuthRequest extends Request {
+  user?: User;
+}
+
 export interface Ctx {
-  req: Request;
+  req: AuthRequest;
   res: Response;
 }
