+++ backend/utils/decorators/roles.decorator.ts
@@ -1,0 +1,3 @@
+import { Reflector } from '@nestjs/core';
+
+export const Roles = Reflector.createDecorator<string[]>();
