+++ backend/utils/decorators/user.decorator.ts
@@ -3,8 +3,9 @@ import { Field, GqlExecutionContext, Int, ObjectType } from '@nestjs/graphql';
 
 export const CurrentUser = createParamDecorator((_data: unknown, context: ExecutionContext) => {
   const ctx = GqlExecutionContext.create(context);
+  const dataFromContext = ctx.getContext().req.user;
 
-  return ctx.getContext().req.user;
+  return dataFromContext;
 });
 
 @ObjectType()
@@ -26,7 +27,4 @@ export class User {
 
   @Field(() => Int)
   group_id: number;
-
-  @Field(() => Boolean)
-  is_admin: boolean;
 }
