+++ frontend/app/[locale]/(admin)/admin/session-admin-provider.tsx
@@ -5,15 +5,15 @@ import { useQuery } from '@tanstack/react-query';
 
 import { fetcher } from '@/graphql/fetcher';
 import {
-  Authorization_Admin_Core_Sessions,
-  Authorization_Admin_Core_SessionsQuery,
-  Authorization_Admin_Core_SessionsQueryVariables
+  Admin_Authorization_Core_Sessions,
+  Admin_Authorization_Core_SessionsQuery,
+  Admin_Authorization_Core_SessionsQueryVariables
 } from '@/graphql/hooks';
 import { SessionAdminContext } from '@/admin/hooks/use-session-admin';
 
 interface Props {
   children: ReactNode;
-  initialDataSession: Authorization_Admin_Core_SessionsQuery | undefined;
+  initialDataSession: Admin_Authorization_Core_SessionsQuery | undefined;
 }
 
 export const SessionAdminProvider = ({ children, initialDataSession }: Props) => {
@@ -23,10 +23,10 @@ export const SessionAdminProvider = ({ children, initialDataSession }: Props) =>
     queryKey: ['Authorization.Admin'],
     queryFn: async () =>
       await fetcher<
-        Authorization_Admin_Core_SessionsQuery,
-        Authorization_Admin_Core_SessionsQueryVariables
+        Admin_Authorization_Core_SessionsQuery,
+        Admin_Authorization_Core_SessionsQueryVariables
       >({
-        query: Authorization_Admin_Core_Sessions
+        query: Admin_Authorization_Core_Sessions
       }),
     initialData: initialDataSession,
     enabled: enableSessionQuery
